requirements.txt
discord.py==2.4.0
yt-dlp==2025.3.1
ffmpeg-python==0.2.0
PyNaCl==1.5.0
lyricsgenius==3.0.1
spotipy==2.24.0
aiohttp==3.10.2
browser-cookie3==0.19.2
requests==2.32.3
python-dotenv==1.0.1